[v0.89] Change freq from int64 to int32, which fixes a compilation issue. Fix Makefile in Windows

+++ src/common/cpu.h
@@ -46,8 +46,8 @@ enum {
 typedef int32_t VENDOR;
 
 struct frequency {
-  int64_t base;
-  int64_t max;
+  int32_t base;
+  int32_t max;
 };
 
 struct hypervisor {
